Modify: Almost the end of the fisrt part
Most of the GET and POST functions are working pretty well.
It needs some functionality to be done, like implement the
lists for the countries and names, and implement the erase
of fields on the database. But for the moment, the work has
been done and can be presented in class.